+++ ProjectWebMVC/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -348,8 +348,8 @@ namespace ProjectWeb.Migrations
                     b.Property<int>("OrderHeaderId")
                         .HasColumnType("int");
 
-                    b.Property<int>("Price")
-                        .HasColumnType("int");
+                    b.Property<double>("Price")
+                        .HasColumnType("float");
 
                     b.Property<int>("ProductId")
                         .HasColumnType("int");
